+++ packages/blocks-core/js/libs/wordpress/read-more/index.js
@@ -17,6 +17,7 @@ import {
  * Internal dependencies
  */
 import type { BlockType } from '../../../type';
+import sharedInnerBlocks from '../inners/shared';
 
 export const ReadMore: BlockType = {
 	name: 'blockeraReadMore',
@@ -32,6 +33,14 @@ export const ReadMore: BlockType = {
 			force: true,
 		},
 		visited: sharedBlockStates.visited,
+	},
+	blockeraInnerBlocks: {
+		'elements/bold': sharedInnerBlocks['elements/bold'],
+		'elements/italic': sharedInnerBlocks['elements/italic'],
+		'elements/kbd': sharedInnerBlocks['elements/kbd'],
+		'elements/code': sharedInnerBlocks['elements/code'],
+		'elements/span': sharedInnerBlocks['elements/span'],
+		'elements/mark': sharedInnerBlocks['elements/mark'],
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;