+++ packages/blocks-core/js/libs/wordpress/read-more/test/functionality.blocks-2.e2e.cy.js
@@ -28,6 +28,15 @@ describe('Read More Block', () => {
 		//
 		cy.checkBlockCardItems(['normal', 'hover', 'focus', 'active']);
 
+		cy.checkBlockStatesPickerItems([
+			'elements/bold',
+			'elements/italic',
+			'elements/kbd',
+			'elements/code',
+			'elements/span',
+			'elements/mark',
+		]);
+
 		//
 		// 3. Edit Block
 		//